+++ examples/analysis/basin_limit_cycle_known.py
@@ -1,5 +1,5 @@
 from dynlib import setup
-from dynlib.analysis import basin_known, ReferenceRun
+from dynlib.analysis import basin_known, basin_axis, ReferenceRun
 from dynlib.plot import export, basin_plot, fig, phase
 
 # Energy Template Oscillator (ETO) with Circular L Curve
@@ -12,8 +12,10 @@ results = basin_known(
     attractors=[
         ReferenceRun(name="Limit Cycle Attractor", ic=[1.0, 0.0]),
     ],
-    ic_grid=[128, 128],
-    ic_bounds=[(-3, 3), (-3, 3)],
+    ic={
+        "x": basin_axis(-3, 3, n=128),
+        "y": basin_axis(-3, 3, n=128),
+    },
     dt_obs=0.01,
     # run long enough to converge + see the cycle
     signature_samples=8000, # samples are in steps not time